main.go: use context to cancel on timeout and improve output
This commit uses context to cancel requests after a timeout that can be
set with flags. It also simplifies the output by removing the
code=[false|true]. False will be displayed with "-" and latencies that
were never probed e.g. one instance was unreachable the fields will
contain empty string "".

Signed-off-by: leonnicolas <leonloechner@gmx.de>